+++ tests/TestCase/Model/Table/AppTableTest.php
@@ -205,7 +205,7 @@ class AppTableTest extends TestCase
         $this->assertFalse($query->valueBinder()->bindings()[':c1']['value']);
 
         $pendingId = collection($query->toArray())->extract('id')->toList();
-        $this->assertEquals([6, 7], $pendingId);
+        $this->assertEquals([6, 8], $pendingId);
     }
 
     /**