+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # 2.x branch
 ## 2.17 branch
+### 2.17.4
+* added `PostsController::getStartAndEndDate()` method;
+* removed `PostsController::indexCompatibility()` method.
+* fixed bug: `PostsTable::getRelated()` returns an empty array if there are no
+    related post;
+* added tests for `PostsCategoriesController` and `PostsController` classes.
+
 ### 2.17.3
 * added `PhotosTable::findPending()` method;
 * fixed all `find()` methods;